fix force update issue, it may block take_queue
when using force_update while this taskid is processing, a new item is
put into task_queue, and overwrite processing.queue_dict[taskid] while
old is not done. This made old one treated as failed after some time.

the second issue is when we change the exetime and priority of item,
    heap is broken. needs resort

the third issue is counter, not counting multiplue times when force
update a active task